+++ src/ValueObject/Point.php
@@ -1,0 +1,20 @@
+<?php declare(strict_types=1);
+
+namespace VysokeSkoly\ImageApi\Sdk\ValueObject;
+
+class Point
+{
+    public function __construct(private readonly int $x, private readonly int $y)
+    {
+    }
+
+    public function getX(): int
+    {
+        return $this->x;
+    }
+
+    public function getY(): int
+    {
+        return $this->y;
+    }
+}