+++ CHANGELOG.md
@@ -2,6 +2,9 @@
 
 <!-- There is always Unreleased section on the top. Subsections (Added, Changed, Fixed, Removed) should be added as needed. -->
 ## Unreleased
+- Add `ImageGenerator` service
+- Add `Crop` Value Object
+- Add `ImageContent::fromImagick` named constructor
 
 ## 3.0.0 - 2022-07-25
 - Require php 8.1